remove unused code

+++ adminapp/views/staff_views.py
@@ -2,48 +2,52 @@ import json
 
 from django.views import generic
 from django.http import HttpResponseRedirect, HttpResponse
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from django.urls import reverse_lazy
 from django.views.generic import View, UpdateView
 from adminapp.models import Users, ProjectStuff
 from adminapp.forms.user_form import UserForm, UserUpdateForm, UserPasswordChangeForm
 from adminapp.views.common_views import CommonView
 from adminapp.views.helper import LogHelper
-from django.conf import settings
 from adminapp.views.project_views import ProjectsView
 
 
 class StaffsView(generic.DetailView):
     def get(self, request):
-        context = CommonView.common_datatable_context(self)
-        return render(request, 'staffs/staff.html', context)
+        if request.user.is_superuser:
+            context = CommonView.common_datatable_context(self)
+            return render(request, 'staffs/staff.html', context)
+        else:
+            return redirect('index')
 
     def delete(request):
         response = {}
-        try:
-            userId = request.POST.get('id')
-            # Users.objects.filter(id=userId).update(is_active='0')
-            Users.objects.get(id=userId).delete()
-            response['success'] = True
-            response['message'] = "User delete successfully"
-        except Exception as e:
-            LogHelper.elog(e)
-            response['success'] = False
-            response['message'] = "Something went wrong. Please try again"
+        if CommonView.superuser_login(request):
+            try:
+                userId = request.POST.get('id')
+                # Users.objects.filter(id=userId).update(is_active='0')
+                Users.objects.get(id=userId).delete()
+                response['success'] = True
+                response['message'] = "User delete successfully"
+            except Exception as e:
+                LogHelper.elog(e)
+                response['success'] = False
+                response['message'] = "Something went wrong. Please try again"
         return HttpResponse(json.dumps(response), content_type='application/json')
 
     def change_user_status(request):
         response = {}
-        try:
-            userId = request.POST.get('id')
-            status = request.POST.get('status')
-            Users.objects.filter(id=userId).update(is_active=status)
-            response['success'] = True
-            response['message'] = "Status Changed Successfully"
-        except Exception as e:
-            LogHelper.elog(e)
-            response['success'] = False
-            response['message'] = "Something went wrong. Please try again"
+        if CommonView.superuser_login(request):
+            try:
+                userId = request.POST.get('id')
+                status = request.POST.get('status')
+                Users.objects.filter(id=userId).update(is_active=status)
+                response['success'] = True
+                response['message'] = "Status Changed Successfully"
+            except Exception as e:
+                LogHelper.elog(e)
+                response['success'] = False
+                response['message'] = "Something went wrong. Please try again"
         return HttpResponse(json.dumps(response), content_type='application/json')
 
 
@@ -52,28 +56,34 @@ class StaffFormView(View):
     template_name = 'staffs/add_staff.html'
 
     def get(self, request, *args, **kwargs):
-        form = self.form_class
-        projects = CommonView.get_all_projects(request)
-        return render(request, self.template_name, {'form': form, 'projects': projects})
+        if request.user.is_superuser:
+            form = self.form_class
+            projects = CommonView.get_all_projects(request)
+            return render(request, self.template_name, {'form': form, 'projects': projects})
+        else:
+            return redirect('index')
 
     def post(self, request, *args, **kwargs):
-        form = self.form_class(request.POST, request.FILES)
-        if form.is_valid():
-            projects = request.POST.getlist('project_list')
-            obj = form.save(self, request)
-            ProjectsView.assign_staff(request, projects, obj.id)
-            mailTemplate = "mails/user_registered.html"
-            context = {
-                "user_full_name": obj.get_full_name(),
-                "password": self.request.POST.get('password'),
-                "username": self.request.POST.get('username')
-            }
-            subject = "Staff Register"
-            to = obj.email
-            CommonView.sendEmail(self.request, mailTemplate, context, subject, to, obj.id)
-            return HttpResponseRedirect('/staffs/')
-        projects = CommonView.get_all_projects(request)
-        return render(request, self.template_name, {'form': form, 'projects': projects})
+        if CommonView.superuser_login(request):
+            form = self.form_class(request.POST, request.FILES)
+            if form.is_valid():
+                projects = request.POST.getlist('project_list')
+                obj = form.save(self, request)
+                ProjectsView.assign_staff(request, projects, obj.id)
+                mailTemplate = "mails/user_registered.html"
+                context = {
+                    "user_full_name": obj.get_full_name(),
+                    "password": self.request.POST.get('password'),
+                    "username": self.request.POST.get('username')
+                }
+                subject = "Staff Register"
+                to = obj.email
+                CommonView.sendEmail(self.request, mailTemplate, context, subject, to, obj.id)
+                return HttpResponseRedirect('/staffs/')
+            projects = CommonView.get_all_projects(request)
+            return render(request, self.template_name, {'form': form, 'projects': projects})
+        else:
+            return redirect('index')
 
 
 class StaffUpdateView(UpdateView):
@@ -83,10 +93,13 @@ class StaffUpdateView(UpdateView):
     form_class = UserUpdateForm
 
     def form_valid(self, form):
-        form.update(request=self.request)
-        projects = self.request.POST.getlist('project_list')
-        ProjectsView.assign_staff(self.request, projects, self.object.id)
-        return HttpResponseRedirect(self.get_success_url())
+        if CommonView.superuser_login(self.request):
+            form.update(request=self.request)
+            projects = self.request.POST.getlist('project_list')
+            ProjectsView.assign_staff(self.request, projects, self.object.id)
+            return HttpResponseRedirect(self.get_success_url())
+        else:
+            return redirect('index')
 
     def get_success_url(self):
         return reverse_lazy('staffs')
@@ -108,14 +121,17 @@ class StaffPasswordChangeView(UpdateView):
         return reverse_lazy('staffs')
 
     def form_valid(self, form):
-        form.save(request=self.request)
-        mailTemplate = "mails/user_password_change.html"
-        context = {
-            "user_full_name": self.object.get_full_name(),
-            "password": self.request.POST.get('password')
-        }
-        subject = "Password Change"
-        to = self.object.email
-        CommonView.sendEmail(self.request, mailTemplate, context, subject, to, self.object.id)
-        return HttpResponseRedirect(self.get_success_url())
+        if CommonView.superuser_login(self.request):
+            form.save(request=self.request)
+            mailTemplate = "mails/user_password_change.html"
+            context = {
+                "user_full_name": self.object.get_full_name(),
+                "password": self.request.POST.get('password')
+            }
+            subject = "Password Change"
+            to = self.object.email
+            CommonView.sendEmail(self.request, mailTemplate, context, subject, to, self.object.id)
+            return HttpResponseRedirect(self.get_success_url())
+        else:
+            return redirect('index')
